Rename sqlite to sql
