+++ scraperwiki/__init__.py
@@ -8,6 +8,4 @@ https://scraperwiki.com/docs/python/python_help_documentation/
 
 from .utils import scrape, pdftoxml, status
 import utils
-import sqlite
-
-sql = sqlite
+import sql
